requirements.txt

pytesseract
Pillow
requests